refactored search to remove searchprovider
